Place postcss-modules plugin first (#249)
Fixes #70 

If the postcss-modules plugin isn't placed first we end up with non-modularized rules along with modularized rules. This situation thus leads to heavier bundles due to code duplication that end up dead and, when the extract option is set to false, it can cause conflicts with other non-modularized (global) rules present in the bundle consumer code.